+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [feat:fix discovery weight param not set to register request bug](https://github.com/Tencent/spring-cloud-tencent/pull/102)
 - [Bugfix: fix causing cpu 100% when set ScheduledThreadPoolExecutor corePoolSize=0](https://github.com/Tencent/spring-cloud-tencent/pull/98)
 - [Feat: optimize router dependency](https://github.com/Tencent/spring-cloud-tencent/pull/110)
+- [Refactor: refactor transfer metadata](https://github.com/Tencent/spring-cloud-tencent/pull/111)